+++ Frends.Pdf.Create/CHANGELOG.md
@@ -1,11 +1,27 @@
 # Changelog
 
+## [2.0.0] - 2026-02-03
+
+### Changed
+
+- [Breaking Change] Upgrade to .net8.0
+- [Breaking Change] Remove deprecated option to choose Unicode vs ANSII text (Code is always Unicode now)
+
+### Added
+
+- Add option to set up custom fonts directory
+- Add option to set up default font
+- Add support for Linux systems
+
+### Fixed
+
+- Resolve issues with not responsive Windows GDI methods (Failing to get fonts)
+
 ## [1.2.0] - 2026-02-03
 
 ### Changed
 
 - Rename task prefix from PDF to Pdf
-
 
 ## [1.1.0] - 2024-08-23